requirements.txt
numpy>=1.23
pandas>=1.5
pyarrow>=12.0
scikit-learn>=1.3
matplotlib>=3.7

pyyaml>=6.0
tqdm>=4.65

jupyter
dvc[gdrive]
ipykernel
python-dotenv
seaborn
pyarrow
fastparquet
openai==1.12.0

tensorflow>=2.12



